+++ Servidor/index.js
@@ -1,15 +1,15 @@
+console.clear();
 const express = require('express');
 require('dotenv').config();
-const bodyParser = require('body-parser');
 const app = express();
+const UserRoueter = require('./src/routes/userRouter.js');
+
+// Middlewares de express
+app.use(express.json());
+app.use(express.urlencoded({ extended: true }));
+
+// Rutas user
+app.use('/user', UserRoueter);
+
 const PORT = process.env.PORT || 3031;
-const UserRoueter = require('../Servidor/src/routes/userRouter.cjs')
-
-app.use(bodyParser.json());
-
-
-app.use('/user',UserRoueter)
-
-
-
 app.listen(PORT, () => console.log(`Servidor escuchando en el puerto http://localhost:${PORT}`));
