--- Servidor/src/config/index.js
@@ -1,1 +1,0 @@
-// configs
