+++ Servidor/src/routes/userRouter.js
@@ -1,0 +1,7 @@
+const express = require('express');
+const { postCreateUser } = require('../controllers/userControle.js');
+const UserRoueter = express.Router();
+
+UserRoueter.post('/crear', postCreateUser);
+
+module.exports = UserRoueter;
